requirements.txt
# Core dependencies
aiogram==3.15.0
aiohttp==3.10.10
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1

psycopg2-binary==2.9.9

# Payment
yookassa==3.1.0

# Image processing
pillow==11.1.0
qrcode==8.0

# Web framework
flask==3.0.3
flask-cors==5.0.0
flask-limiter==3.8.0
flask-talisman==1.1.0
gunicorn==21.2.0

# Utilities
requests==2.32.3
cryptography==44.0.0
